[WIP] Change main method to `cql.recommend` to prevent mixing up function with namespace
