+++ src/cql.ts
@@ -1,6 +1,5 @@
 export const version = '__VERSION__';
 
-import {extend} from './util';
 
 export import config = require('./config');
 export import constraint = require('./constraint/constraint');
@@ -13,10 +12,8 @@ export import nest = require('./nest');
 export import property = require('./property');
 
 // Make it so that we can call cql.query() as method, or access other methods inside cql.query
-import * as cqlQuery from './query/query';
-import {query as queryFn} from './query/query';
-
-export const query = extend(queryFn, cqlQuery);
+export import query = require('./query');
+export {recommend} from './recommend';
 
 export import ranking = require('./ranking/ranking');
 export import schema = require('./schema');
