+++ src/query/index.ts
@@ -1,41 +1,14 @@
-import {Config} from 'vega-lite/src/config';
-
 import {SpecQuery} from './spec';
 import {GroupBy} from './groupby';
 
-import {QueryConfig, DEFAULT_QUERY_CONFIG} from '../config';
-import {generate} from '../generate';
-import {nest} from '../nest';
-import {rank} from '../ranking/ranking';
-import {Schema} from '../schema';
-import {duplicate, extend} from '../util';
-
-import {SpecQueryModelGroup} from '../model';
+import {QueryConfig} from '../config';
+import {duplicate} from '../util';
 
 export import encoding = require('./encoding');
 export import groupBy = require('./groupby');
 export import shorthand = require('./shorthand');
 export import spec = require('./spec');
 export import transform = require('./transform');
-
-export function query(q: Query, schema: Schema, config?: Config): {query: Query, result: SpecQueryModelGroup} {
-  // 1. Normalize non-nested `groupBy` to always have `groupBy` inside `nest`
-  //    and merge config with the following precedence
-  //    query.config > config > DEFAULT_QUERY_CONFIG
-  q = extend({}, normalize(q), {
-    config: extend({}, DEFAULT_QUERY_CONFIG, config, q.config)
-  });
-
-  // 2. Generate
-  const answerSet = generate(q.spec, schema, q.config);
-  const nestedAnswerSet = nest(answerSet, q);
-  const result = rank(nestedAnswerSet, q, schema, 0);
-
-  return {
-    query: q,
-    result: result
-  };
-}
 
 /**
  * Normalize the non-nested version of the query to a standardize nested
