+++ test/nest.test.ts
@@ -10,7 +10,7 @@ import {generate} from '../src/generate';
 import {SpecQueryModel, SpecQueryModelGroup} from '../src/model';
 import {nest, FIELD, FIELD_TRANSFORM, ENCODING, TRANSPOSE} from '../src/nest';
 import {Property} from '../src/property';
-import {Query} from '../src/query/query';
+import {Query} from '../src/query';
 import {REPLACE_BLANK_FIELDS, REPLACE_XY_CHANNELS, REPLACE_FACET_CHANNELS, REPLACE_MARK_STYLE_CHANNELS} from '../src/query/groupby';
 
 import {SHORT_WILDCARD} from '../src/wildcard';
